Change isize/usize args & fields to i32/u32

+++ src/label.rs
@@ -52,7 +52,7 @@ impl Click for Label {
 }
 
 impl Place for Label {
-    fn position(self, x: isize, y: isize) -> Self {
+    fn position(self, x: i32, y: i32) -> Self {
         let mut rect = self.rect.get();
         rect.x = x;
         rect.y = y;
@@ -61,7 +61,7 @@ impl Place for Label {
         self
     }
 
-    fn size(self, width: usize, height: usize) -> Self {
+    fn size(self, width: u32, height: u32) -> Self {
         let mut rect = self.rect.get();
         rect.width = width;
         rect.height = height;
@@ -85,7 +85,7 @@ impl Widget for Label {
                 x = 0;
                 y += 16;
             }else{
-                if x + 8 <= rect.width as isize && y + 16 <= rect.height as isize {
+                if x + 8 <= rect.width as i32 && y + 16 <= rect.height as i32 {
                     renderer.char(Point::new(x, y) + rect.point(), c, self.fg);
                 }
                 x += 8;
